chore(team): split runtime view manager tests
